+++ integration/graphql/src/main/java/com/blazebit/persistence/integration/graphql/GraphQLDefaultFetch.java
@@ -7,6 +7,7 @@ package com.blazebit.persistence.integration.graphql;
 
 import com.blazebit.persistence.view.EntityViewSetting;
 import java.lang.annotation.ElementType;
+import java.lang.annotation.Repeatable;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
@@ -22,6 +23,7 @@ import java.lang.annotation.Target;
  */
 @Target({ElementType.METHOD, ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
+@Repeatable(GraphQLDefaultFetches.class)
 public @interface GraphQLDefaultFetch {
     /**
      * Specifies the GraphQL field name that has to be present in the selection set to enable default fetching